process {multicore}	R Documentation

Function to query objects of the class process

Description

processID returns the process IDs for the given processes. It raises an error if process is not an object of the class process or a list of such objects.

print methods shows the process ID and its class name.

Usage

processID(process)
## S3 method for class 'process'
print(x, ...)

Arguments

process
process (object of the class process) or a list of such objects.

x
process to print

...
ignored

Value

processID returns an integer vector contatining the process IDs.

print returns NULL invisibly

Author(s)

Simon Urbanek

See Also

fork

[Package multicore version 0.1-7 Index]