sendMaster {multicore}	R Documentation

Sends data from the child to to the master process

Description

sendMaster Sends data from the child to to the master process

Usage

sendMaster(what)

Arguments

what
data to send to the master process. If what is not a raw vetor, what will be serialized into a raw vector. Do NOT send an empty raw vector - it is reserved for internal use.

Details

Any child process (created by fork directly or by parallel indirectly) can send data to the parent (master) process. Usually this is used to deliver results from the parallel child processes to the master process.

Value

returns TRUE

Author(s)

Simon Urbanek

See Also

parallel, fork

[Package multicore version 0.1-7 Index]