signals {multicore}	R Documentation

Signal constants (subset)

Description

SIGALRM alarm clock

SIGCHLD to parent on child stop or exit

SIGHUP hangup

SIGINFO information request

SIGINT interrupt

SIGKILL kill (cannot be caught or ignored)

SIGQUIT quit

SIGSTOP sendable stop signal not from tty

SIGTERM software termination signal from kill

SIGUSR1 user defined signal 1

SIGUSR2 user defined signal 2

Details

See man signal in your OS for details. The above codes can be used in conjunction with the kill function to send signals to processes.

Author(s)

Simon Urbanek

See Also

kill

[Package multicore version 0.1-7 Index]